feat: refactore elements add webpack
