+++ index.html
@@ -29,6 +29,6 @@
       </section>
     </div>
 
-    <script src="js/script.js"></script>
+    <script src="js/bundle.js"></script>
   </body>
 </html>
