+++ webpack.config.js
@@ -1,0 +1,13 @@
+'use strict';
+
+let path = require('path');
+
+module.exports = {
+  mode: 'development',
+  entry: './js/script.js',
+  output: {
+    filename: 'bundle.js',
+    path: __dirname + '/js',
+  },
+  watch: true,
+};
